requirements.txt
Flask==0.10.1
GitPython==0.3.2.RC1
Markdown==2.3.1
